pyproject.toml
[tool.poetry]
name = "term-paper"
version = "0.1.0"
description = ""
authors = ["waron97 <winkler.aron5gmail.com>"]
readme = "README.md"
packages = [{include = "term_paper"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.7"
requests = "^2.28.2"
dpath = "^2.1.5"
tqdm = "^4.65.0"
xmltodict = "^0.13.0"
nltk = "^3.8.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
flake8 = "^6.0.0"
pylance = "^0.3.18"
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
